proc: initial work

+++ kernel/src/mm/alloc.cc
@@ -1,7 +1,7 @@
+#include <cstddef>
 #include <internal/lock.h>
 #include <mrk/pmm.h>
 #include <mrk/vmm.h>
-#include <cstddef>
 
 static mutex alloc_lock;
 
@@ -9,41 +9,39 @@ extern "C" {
 
 int liballoc_lock()
 {
-	alloc_lock.lock();
-	return 0;
+    alloc_lock.lock();
+    return 0;
 }
 
 int liballoc_unlock()
 {
-	alloc_lock.unlock();
-	return 0;
+    alloc_lock.unlock();
+    return 0;
 }
 
 void* liballoc_alloc(int pages)
 {
-	uint64_t pgs = (uint64_t)mm::pmm::get(pages);
-	void* to_ret = (void*)(pgs + MEM_VIRT_OFFSET);
+    uint64_t pgs = (uint64_t)mm::pmm::get(pages);
+    void* to_ret = (void*)(pgs + MEM_VIRT_OFFSET);
 
-	for(size_t i = 0; i < pages; i++) {
-            mm::vmm::kernel_space()->map(pgs, pgs + MEM_VIRT_OFFSET, PAGE_PRESENT | PAGE_WRITEABLE | PAGE_NX | PAGE_GLOBAL);
-	    pgs += 0x1000;
-	}
+    for (int i = 0; i < pages; i++) {
+        mm::vmm::kernel_space()->map(pgs, pgs + MEM_VIRT_OFFSET, PAGE_PRESENT | PAGE_WRITEABLE | PAGE_NX | PAGE_GLOBAL);
+        pgs += 0x1000;
+    }
 
-	return to_ret;
+    return to_ret;
 }
 
 int liballoc_free(void* ptr, int pages)
 {
-	mm::pmm::free((void*)((uint64_t)ptr - MEM_VIRT_OFFSET), pages);
+    mm::pmm::free((void*)((uint64_t)ptr - MEM_VIRT_OFFSET), pages);
 
-	uint64_t pgs = (uint64_t)ptr;
-	for(size_t i = 0; i < pages; i++) {
-            mm::vmm::kernel_space()->unmap(pgs);
-            pgs += 0x1000;
-        }
+    uint64_t pgs = (uint64_t)ptr;
+    for (int i = 0; i < pages; i++) {
+        mm::vmm::kernel_space()->unmap(pgs);
+        pgs += 0x1000;
+    }
 
-	return 0;
+    return 0;
 }
-
 }
-
